updated connector types component
