+++ app/cdap/components/ConnectorTypes/ConnectorTypeCard/types.ts
@@ -1,0 +1,4 @@
+export interface IConnectorTypeCardComponentProps {
+  name: string;
+  image: JSX.Element;
+}
